Refactored Pick File into CH<x>.py files.
