+++ Code Work-Julion/chclass.py
@@ -1,3 +1,6 @@
+import os.path
+import array
+import math
 import random
 import pathlib
 
@@ -19,7 +22,7 @@ def classchooser():
 
 def classrollertry(ClassChoosen):
     while True:
-        classtest = ClassChoosen
+        classtest = str(ClassChoosen)
         opentest = False
         classpath = "CharDatabase/Class/"+str(classtest)+".txt"
         path = pathlib.Path(str(classpath))
@@ -51,7 +54,6 @@ def archchoose(classname):
         return str(gotarch)
     except:
         print("Somehow an error in the archetype choosing!!")
-    return str(gotarch)
 
 
 def fullclass():
@@ -64,18 +66,54 @@ def fullclass():
         print("Current Class: "+str(arch)+" "+str(myClass))
         classFlag = False
         while True:
-            incheck = input("Reroll? (y/n): ")
+            print("Yes / No / Choose My Own")
+            incheck = input("Reroll? (y/n/c): ")
             if incheck == "n":
                 classFlag = True
                 break
             if incheck == "y":
                 classFlag = False
                 break
+            if incheck == "c":
+                myClass = pickclass()
+                arch = archchoose(str(myClass))
+                while True:
+                    choosecheck = input("You have gotten " + str(arch) + " " + str(myClass) + ". Keep? (y/n): ")
+                    if choosecheck == "y":
+                        return myClass, arch
+                    if choosecheck == "n":
+                        break
+                    else:
+                        print("Only 'y' or 'n' please.")
             else:
                 print("Not Correct. 'y' or 'n' Only Please. ")
         if classFlag is True:
             break
     return myClass, arch
+
+def pickclass():
+    try:
+        clList = open("CharDatabase/Class/$List.txt")
+        clread = clList.read()
+        clEntries = clread.split("\n")
+        clnum = len(clEntries)
+        clTrueList = []
+        for x in clEntries:
+            if classrollertry(str(x)):
+                clTrueList.append(str(x))
+        i = 0
+        for y in clTrueList:
+            print("Number ["+str(i)+"]: "+str(y))
+            i = i+1
+        while True:
+            num = input("Enter Number You Want: ")
+            if (int(num) <= i-1) and (int(num) >= 0):
+                print("Chosen: " + str(clTrueList[int(num)]))
+                return clTrueList[int(num)]
+            else:
+                print("Try Again.")
+    finally:
+        clList.close()
 
 def classbaseadd(filename, myclass, level):
     try:
